// src/tests/workEmail/validateWorkEmail.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type ValidateWorkEmailMutationVariables = Types.Exact<{
  confirmationToken: Types.Scalars['String']['input'];
}>;


export type ValidateWorkEmailMutation = { __typename?: 'Mutation', validateWorkEmail: { __typename: 'WorkEmail', id: string, isValidated: boolean } };


export const ValidateWorkEmail = gql`
    mutation ValidateWorkEmail($confirmationToken: String!) {
  validateWorkEmail(confirmationToken: $confirmationToken) {
    __typename
    id
    isValidated
  }
}
    `;